The test plots now go to /dev/null
I don't actually want to see the test plots, I just want to make sure that no
errors are thrown. I thus send the output to /dev/null. Note that I never
actually read stdout, so if this test plot goes to the default stdout output,
then eventually the buffer fills up and gnuplot blocks. So keep it going to
/dev/null, or make sure to read the test plot from stdout