plot retains bounds, if unspecified
If a plot() does not set a bound, then we do not reset it. The old behavior
would set a default bound if it was unset. This was highly inconvenient in the
case where the user would zoom in on something, then replot a slightly different
dataset, and then lose their focus area