'with labels' plots work now
I thought these were functional already, but I guess not fully. These worked if
passing in numerical data, and printing the numbers as the labels. But plotting
generic text as my labels was broken. It now works to plot labels at arbitrary
locations. For instance:

    x = np.arange(3)
    y = x*x
    gp.plot(y, np.array(('aaa', 'bx xb', 'ccc')), _with='labels', tuplesize=3, ascii=1)

Matrix 'with labels' plots don't work, but that's a gnuplot bug I think.