add_plot_option() takes multiple values in a list
This is a bit uglier, but it has exactly the same semantics as the plot options
that can take multiple values. I can now

  gp.add_plot_option(opts0, 'set', opts1['set'])

Previously I would need to do either that or

  gp.add_plot_option(opts0, 'set', *opts1['set'])

depending on what was actually in opts1['set'].

THIS IS AN API BREAK. But this feature is less than one week old, so I think
it's ok. Nobody is using it